+++ app/pit-survey/page.js
@@ -3,7 +3,7 @@ import Image from "next/image";
 import { Button, Autocomplete, FormControl, FormLabel, Input, RadioGroup, Radio, List, ListItem, Checkbox, FormHelperText, Snackbar, Textarea } from '@mui/joy'
 import MenuButton from "@/components/menu-button";
 import { SFLAllTeams } from "../data/sfl-all-teams";
-import { useState, useRef } from "react";
+import { useState, useRef, useDebugValue } from "react";
 import styles from './pit.module.css'
 
 //images
@@ -11,6 +11,8 @@ import WCD from '../../public/images/westcoastdrive.png'
 import Mec from '../../public/images/mecanumdrive.png'
 import Tank from '../../public/images/tankdrive.jpg'
 import Swerve from '../../public/images/swervedrive.jpg'
+
+// import type { PutBlobResult } from '@vercel/blob'; //js quickstart does not have typing but has defined w `as`
 
 export default function PitSurveyPage() {
   const [teamNumber, setTeamNumber] = useState('')
@@ -23,12 +25,14 @@ export default function PitSurveyPage() {
   const [helpClimb, setHelpClimb] = useState(null)
   const [scoreClimb, setScoreClimb] = useState(null)
   const [investigate, setInvestigate] = useState('')
-  const [feedback, setFeedback] = useState(null)
+  const [feedback, setFeedback] = useState('')
   const [name, setName] = useState('')
 
   //form state
   const [loading, setLoading] = useState(false)
   const formRef = useRef(null);
+  const [blob, setBlob] = useState(null)
+  var instantlyKnowIfSubmit = false;  //stupid name lol but used bc state is async
   // const [validated, setValidated] = useState(false)
 
   //checkboxes
@@ -43,10 +47,19 @@ export default function PitSurveyPage() {
   const [swerveSelected, setSwerve] = useState(false)
   const [isOtherSelected, setOther] = useState(false)
 
+  // const [frontImage, setFrontImage] = useState()
+  // const [sideImage, setSideImage] = useState()
+  const frontImageRef = useRef()
+  const sideImageRef = useRef()
+  const [frontImageSize, setFrontImageSize] = useState(0.0)
+  const [sideImageSize, setSideImageSize] = useState(0.0)
+  const [unit, setUnit] = useState('MB')
+
   //snackbar state
   const [open, setOpen] = useState(false)
   const [errorString, setErrorString] =useState('')
   const [submitSuccess, setSuccess] = useState(false)
+  const [color, setColor] = useState('neutral')
 
   function handleCheckbox(value, checked){
     console.log(`${value}, ${checked}`)
@@ -118,10 +131,86 @@ export default function PitSurveyPage() {
     }
   }
 
+  //grab first element from target because this is
+  //meant to be a array capable of holding multiple images if needed
+  function handleImages(e){
+    const eventTarget = e.target
+    const file = eventTarget.files[0]
+    //show file size bc Vercel Blob only allows a Server Upload of 4.5 MB
+    // const size = file.size / (1024 * 1024).toFixed(2)
+    const frontPreview = document.getElementById('preview-1')
+    const sidePreview = document.getElementById('preview-2')
+
+    const bytes = file.size
+    // console.log(bytes)
+    var convertedSize = 0.0;
+    if(bytes < 1000000){
+      setUnit('KB')
+      convertedSize = Math.floor(bytes/1000).toFixed(2);
+    } else{
+        setUnit('MB')
+        convertedSize = Math.floor(bytes/1000000).toFixed(2); 
+    }
+
+    if(eventTarget.id === 'front-picture'){
+      try {
+        // setFrontImage(file)
+        frontImageRef.current = file
+        console.log(frontImageRef.current)
+        setFrontImageSize(convertedSize)
+        setColor('primary')
+        setErrorString('Attached front image!')
+
+        const img = document.createElement("img");
+        img.classList.add("preview-img");
+        img.file = file;
+        frontPreview.appendChild(img);
+    
+        const reader = new FileReader();
+        reader.onload = (e) => {
+          img.src = e.target.result;
+        };
+        reader.readAsDataURL(file);
+      } catch (error) {
+        setColor('warning')
+        setErrorString('Error attaching front image!')
+        alert(error)
+      }
+    }
+    else if(eventTarget.id === 'side-picture'){
+      try {
+        // setSideImage(file)
+        sideImageRef.current = file
+        console.log(sideImageRef.current)
+        setSideImageSize(convertedSize)
+        setColor('primary')
+        setErrorString('Attached side image!')  //custom string aside from vanilla "submitted!"
+
+        const img = document.createElement("img");
+        img.classList.add("preview-img");
+        img.file = file;
+        sidePreview.appendChild(img);
+    
+        const reader = new FileReader();
+        reader.onload = (e) => {
+          img.src = e.target.result;
+        };
+        reader.readAsDataURL(file);
+      } catch (error) {
+        setColor('warning')
+        setErrorString('Error attaching side image!')
+        alert(error)
+      }
+    }
+
+    setOpen(true) //show snackbar after color/string has been set
+  }
+
   function handleValidate(passedEvent){
     if (teamNumber==='' || !drivetrain || !prefPos || !vision
       || !scoreHeight || !pickup || !climb || !investigate || !name){
     setErrorString('All fields required!')
+    setColor('danger')
     setSuccess(false)
     setOpen(true)
     return false;
@@ -129,6 +218,7 @@ export default function PitSurveyPage() {
 
     else if(feedback && feedback.length > 255){
       setErrorString('Feedback must be at most 255 chars.')
+      setColor('danger')
       setSuccess(false)
       setOpen(true)
       return false;
@@ -137,7 +227,7 @@ export default function PitSurveyPage() {
     handleSubmit(passedEvent)
   }
 
-  function handleSubmit(e){
+  const handleSubmit = async (e) => {
     setLoading(true)
     e.preventDefault()
 
@@ -170,6 +260,7 @@ export default function PitSurveyPage() {
     .then((response => {
       if(!response.ok){
           setSuccess(false)
+          setColor('danger')
 
           switch (response.status) {
             case 400:
@@ -185,7 +276,19 @@ export default function PitSurveyPage() {
           }
       } else { //reset
           setSuccess(true)
+          setColor('success')
+          instantlyKnowIfSubmit = true;
+          console.log(instantlyKnowIfSubmit)
+          console.log(frontImageRef.current)
+          console.log(sideImageRef.current)
+          uploadImages()
+
           formRef.current.reset();
+
+          frontImageRef.current = null;
+          sideImageRef.current = null; //reset after submitting
+          console.log(frontImageRef.current)
+          console.log(sideImageRef.current)
 
           setTeamNumber('')
 
@@ -208,16 +311,64 @@ export default function PitSurveyPage() {
           setHelpClimb(null)
           setScoreClimb(null)
           setInvestigate('')
-          setFeedback(null)
+          setFeedback('')
           setName('')
+
+          setFrontImageSize(0.0)
+          setSideImageSize(0.0)
+
+          document.querySelectorAll('.preview-img').forEach((img) => {
+              img.remove();
+            });
+          console.log('preview-1`s img removed')
+          console.log('preview-2`s img removed')
       }
       setOpen(true)
       setLoading(false)
   }))
   .catch(error => {
-      console.log(error.json())
+      console.log(error)
   })
+
+  console.log('outside submit')
+}
+
+async function uploadImages(){
+  if(instantlyKnowIfSubmit == true){ // await only allowed at upper level so wrap in conditional
+    console.log('reached img upload')
+    const front = frontImageRef.current
+    const side = sideImageRef.current
+    const imageData = {
+      front: front,
+      side: side
+    }
+
+    setLoading(true)
+    setColor('neutral')
+    setErrorString("Uploading images...")
+    setSuccess(false)
+    setOpen(true)
+
+    const response = await fetch(
+      `/api/upload-pit-images?filename_front=${front.name}&filename_side=${side.name}`,
+      {
+        method: 'POST',
+        body: JSON.stringify(imageData),
+      },
+    ).then((response => {
+      if(!response.ok){
+        setSuccess(false)
+        setColor('danger')
+        setErrorString("Error uploading images!")
+      } else {
+        const newBlob = (response.json()) // as PutBlobResult;
+        console.log(newBlob)
+
+        setBlob(newBlob);
+      }
+    }));
   }
+}
 
   return (
     <>
@@ -231,7 +382,7 @@ export default function PitSurveyPage() {
           <Autocomplete
             required
             type="number"
-            inputMode="numeric"
+            inputMode="tel"
             options={SFLAllTeams}
             value={teamNumber}
             onChange={handleInputChange}
@@ -369,6 +520,37 @@ export default function PitSurveyPage() {
           </>
         )}
 
+        <h2 style={{marginBottom:0}}>Pictures</h2>
+        <small>Max size of each image: 4.5 MB</small>
+        <FormControl  sx={{ marginBottom: '1rem', height:'fit-content !important'}}>
+          <label htmlFor="front-picture"><strong>Front</strong> View:</label>
+          <input
+            type="file"
+            ref={frontImageRef}
+            id="front-picture"
+            name="picture"
+            accept="image/*"
+            capture="environment" //! this is what allows for camera functionality on mobile. desktop triggers file browser
+            onChange={handleImages}
+          />
+          <output id='filesize-front'><small>{frontImageSize} {unit}</small></output>
+          <div id="preview-1" className={styles.imgPreview}></div>
+        </FormControl>
+        
+        <FormControl  sx={{ marginBottom: '1rem', height:'fit-content !important'}}>
+          <label htmlFor="side-picture"><strong>Side</strong> View:</label>
+          <input
+            type="file"
+            ref={sideImageRef}
+            id="side-picture"
+            name="picture"
+            accept="image/*"
+            capture="environment"  //! this is what allows for camera functionality on mobile. desktop triggers file browser
+            onChange={handleImages}/>
+          <output id='filesize-side'><small>{sideImageSize} {unit}</small></output>
+          <div id="preview-2" className={styles.imgPreview}></div>
+        </FormControl>
+
         <h2>Information</h2>
         <FormControl  sx={{ marginBottom: '1rem'}}>
           <FormLabel>Do you think this robot is worth investigating? <sup className='req'>*</sup></FormLabel>
@@ -407,8 +589,8 @@ export default function PitSurveyPage() {
 
         <Snackbar
         variant="solid"
-        color={submitSuccess ? 'success' : 'danger'}
-        autoHideDuration={submitSuccess ? 3500 : 5000}
+        color={color}
+        autoHideDuration={errorString === 'Uploading images...' ? null : (submitSuccess ? 3500 : 5000 ?? 3500)}
         open={open}
         onClose={() => setOpen(false)}
         // onUnmount={handleReset}
@@ -417,8 +599,13 @@ export default function PitSurveyPage() {
         `Submitted!`
         : `${errorString}`}
         </Snackbar>
+
+        {/* testing only */}
+        {blob && (
+        <div>
+          Blob url: <a href={blob.url}>{blob.url}</a>
+        </div>
+      )}
     </>
-
-
   )
 }