+++ next.config.mjs
@@ -1,6 +1,15 @@
 /** @type {import('next').NextConfig} */
 const nextConfig = {
-    reactStrictMode: false
+    reactStrictMode: false,
+    images: {
+        remotePatterns: [
+          {
+            protocol: 'https',
+            hostname: 'pit-images.public.blob.vercel-storage.com',
+            port: '',
+          },
+        ],
+      }
 };
 
 export default nextConfig;